requirements.txt
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
idna==2.8
kiwisolver==1.1.0
matplotlib==3.0.3
numpy==1.16.3
pandas==0.24.2
pymongo==3.8.0
pyparsing==2.4.0
python-dateutil==2.8.0
python-dotenv==0.10.2
pytz==2019.1
requests==2.21.0
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
urllib3==1.24.3
